Add a block_device_apply method to the udev module
Use where appropriate.

Entirely remove the identify() method but substitute applications of
block_device_apply where appropriate in testing.
