+++ src/engine/strat_engine/udev.rs
@@ -3,7 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 //! udev-related methods
-use std::ffi::OsStr;
+use std::{ffi::OsStr, path::Path};
 
 use libudev;
 
@@ -116,4 +116,27 @@ pub fn decide_ownership(device: &libudev::Device) -> StratisResult<UdevOwnership
     } else {
         UdevOwnership::Theirs
     })
+}
+
+/// Locate a udev block device with the specified devnode and apply a function
+/// to that device, returning the result.
+/// This approach is necessitated by the libudev lifetimes, which do not allow
+/// returning anything directly obtained from the enumerator value created in
+/// the method itself.
+/// Note that this does require iterating through the blockdevs in the udev
+/// database, so it is essentially linear in the number of block devices.
+pub fn block_device_apply<F, U>(devnode: &Path, f: F) -> StratisResult<Option<U>>
+where
+    F: FnOnce(&libudev::Device) -> U,
+{
+    let canonical = devnode.canonicalize()?;
+
+    let context = libudev::Context::new()?;
+    let mut enumerator = block_enumerator(&context)?;
+
+    Ok(enumerator
+        .scan_devices()?
+        .filter(|dev| dev.is_initialized())
+        .find(|x| x.devnode().map_or(false, |d| canonical == d))
+        .map(|d| f(&d)))
 }
